Latest modifs from xenbits

+++ src/xencrsh/hvm.c
@@ -1,4 +1,5 @@
-/* Copyright (c) Citrix Systems Inc.
+/* Copyright (c) Xen Project.
+ * Copyright (c) Cloud Software Group, Inc.
  * All rights reserved.
  * 
  * Redistribution and use in source and binary forms, 
@@ -30,8 +31,8 @@
  */ 
 
 #include <wdm.h>
-#include <xenvbd-storport.h>
 #include <intrin.h>
+#include <xenvbd-storport.h>
 
 #include <xen-version.h>
 #include <xen\xen-compat.h>
@@ -103,21 +104,21 @@ CpuId(
     OUT PULONG  EDX OPTIONAL
     )
 {
-    int       Value[4] = {0};
+    int         Value[4] = {0};
 
     __cpuid(Value, Leaf);
 
     if (EAX)
-        *EAX = Value[0];
+        *EAX = (ULONG)Value[0];
 
     if (EBX)
-        *EBX = Value[1];
+        *EBX = (ULONG)Value[1];
 
     if (ECX)
-        *ECX = Value[2];
+        *ECX = (ULONG)Value[2];
 
     if (EDX)
-        *EDX = Value[3];
+        *EDX = (ULONG)Value[3];
 }
 
 static FORCEINLINE NTSTATUS
