+++ src/xenvbd/driver.c
@@ -1,4 +1,5 @@
-/* Copyright (c) Citrix Systems Inc.
+/* Copyright (c) Xen Project.
+ * Copyright (c) Cloud Software Group, Inc.
  * All rights reserved.
  * 
  * Redistribution and use in source and binary forms, 
@@ -301,7 +302,7 @@ DriverGetFeatureOverride(
 {
     BOOLEAN              Present = FALSE;
 
-    if (Feature < ARRAYSIZE(Driver.FeatureOverride)) {
+    if ((ULONG)Feature < ARRAYSIZE(Driver.FeatureOverride)) {
         Present = Driver.FeatureOverride[Feature].Present;
         *Value = Driver.FeatureOverride[Feature].Value;
     }
@@ -315,7 +316,7 @@ DriverGetFeatureName(
     IN  XENVBD_FEATURE  Feature
     )
 {
-    return (Feature < ARRAYSIZE(Driver.FeatureOverride)) ?
+    return ((ULONG)Feature < ARRAYSIZE(Driver.FeatureOverride)) ?
            Driver.FeatureOverride[Feature].Name :
            NULL;
 }
